Add category-specific approval and setting approver by page property
Allows specifying $ApprovedRevsCategoryApprovals to define categories that
are able to be approved by people in specified groups, people specified by
a page property (set by {{#approvedrevs_approvers: Some | Approver|User }}
and anyone with right 'approverevisions'. Example configuration:

$egApprovedRevsCategoryApprovals = [
  'Some_Category' => [
    'group' => ['Approvers','SomeOtherGroup'],
    'property' => true
  ],
  'Approved_Revision_Required' => [
    'group' => 'ImportantPeople',
    'property' => false
  ],
];

This would allow pages in Some_Category to be approved by people in groups
'Approvers" and 'SomeOtherGroup', as well as people specified by the
parser function. Category Approved_Revision_Required would only be able to
be approved by people in group 'ImportantPeople', and not by people set
via parser function. Both categories would be able to be approved by
anyone with right 'approverevisions' (default: sysop).

This could also be used along with $ApprovedRevsNamespaces. Not all users
may be able to approve in NS_MAIN, but this could allow them to approve on
specific categories. Likewise, Use of {{#approvedrevs_aprovers: ...}} does
not need to be used in conjunction with a category. It could be used on a
page within a $ApprovedRevsNamespaces page to add more approvers.

Change-Id: I5e9b791da7e4bd4e83c20e96b471b890006cf278